code cleanup; additional security checks

+++ admin/item_delete.php
@@ -4,42 +4,35 @@ session_start();
 
 require_once("../functions.php");
 
+// Redirect to login page if the user is not logged in
 if (!isset($_SESSION['user'])) {
-	header("location:login.php");
+    header("Location: login.php");
+    exit();
 }
 
-if (isset($_POST['submit'])) {
+$statusMessage = '';
 
-	$statusMessage = deleteItem();
-
+if ($_SERVER['REQUEST_METHOD'] === 'POST' && isset($_POST['submit'])) {
+    $statusMessage = deleteItem();
 }
-
 
 require_once("header.php");
 
-if (isset($statusMessage)) {
-	echo "<span style='color: green;'>$statusMessage</span><br>";
+if (!empty($statusMessage)) {
+    echo "<span style='color: green;'>$statusMessage</span><br>";
 }
+?>
 
-echo '
+<form action="<?php echo htmlspecialchars($_SERVER['PHP_SELF']); ?>" method="POST">
+    <div class="form-group">
+        <p>&nbsp;</p>
+        <?php getAllItemsForDelete(); ?>
+        <input type="checkbox" id="deleteme" name="deleteme" value="1">&nbsp;
+        <label for="deleteme">Yes, I am sure!</label>
+    </div>
+    <input class="btn btn-primary" type="submit" name="submit" value="Delete">
+</form>
 
-		<form action="'.$_SERVER['PHP_SELF'].'" method="POST">
-
-			<div class="form-group">
-			<p>&nbsp;</p>';
-
-			getAllItemsForDelete();
-
-echo '	
-				<input type="checkbox" id="deleteme" name="deleteme" value="1">&nbsp;
-				<label for="deleteme">Yes I am sure!</label>
-			</div>
-
-			<input class="btn btn-primary" type="submit" name="submit" value="Delete">
-
-		</form>
-';
-
+<?php
 require_once("footer.php");
-
 ?>
